test case for sentinel master listener
